Library Database
The Library Management System project creates a database for a
librarians use in a library. The librarian can SEARCH, CHECK OUT and
CHECK IN books as well as add a new BORROWER and calculate the FINE for
books that have not been turned in on or before the due date.